+++ CHANGELOG.md
@@ -7,10 +7,18 @@ All notable changes to this project will be documented in this file.
 ### Added 
 
 - Support for specifying absolute/relative path to app executable file instead of app_dir and relative app file separately
+- Log files - option to append or overwrite files.
 
 ### Fixed
 
-- Log files - option to append or overwrite files.
+- Add getters/setters to UniversalBatchExecuter class.
+
+## [0.0.4] - 2021-02-16
+
+### Fixed
+
+- Error message shwowing when incorrect data in config.json
+- Some minor changes to the project structure.
 
 ## [0.0.3] - 2021-02-15